@@ -7,9 +7,9 @@ import android.hardware.Camera
 import android.net.Uri
 import android.os.Bundle
 import android.os.Environment
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.FileProvider
-import android.support.v7.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.FileProvider
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import android.view.animation.AccelerateInterpolator
 import android.view.animation.Animation
@@ -152,7 +152,7 @@ class CameraActivity : AppCompatActivity(), Camera.PictureCallback {
                 preferencesHelper.phoneId,
                 preferencesHelper.phoneName
             ))
-        FileUploadService.enqueueWork(this, intent)
+        startForegroundService(intent)
     }
 
     private fun onError() {
