@@ -18,7 +18,7 @@ class CameraPreview(
         holder.addCallback(this)
     }
 
-    override fun surfaceChanged(holder: SurfaceHolder?, format: Int, width: Int, height: Int) {
+    override fun surfaceChanged(holder: SurfaceHolder, format: Int, width: Int, height: Int) {
         if (this.holder.surface == null) {
             return
         }
@@ -47,11 +47,11 @@ class CameraPreview(
         camera.startPreview()
     }
 
-    override fun surfaceDestroyed(holder: SurfaceHolder?) {
+    override fun surfaceDestroyed(holder: SurfaceHolder) {
 
     }
 
-    override fun surfaceCreated(holder: SurfaceHolder?) {
+    override fun surfaceCreated(holder: SurfaceHolder) {
         try {
             camera.setPreviewDisplay(holder)
         } catch(e: RuntimeException) {
